Implements tutorial context switching
Adds tutorial context switching to all using E, Space or Enter to continue during tutorials and end of mining.

This change introduces GUIDE context switching to the tutorial sequences in Home Base, Mining Area, and End Mining screens. It sets the tutorial context when the tutorial is active and restores the main context when it's finished.